solutions/ehc/web/ehc_web/ehc/api/requirements.txt
djangorestframework==3.3.1
Django==1.8.5
Pygments==2.0.2
#Markdown==2.6.3

# heroku
# run these commands to setup heroku properly
# heroku create --stack cedar --buildpack git://github.com/heroku/heroku-buildpack-python.git
# heroku config:add BUILDPACK_URL=git@github.com:heroku/heroku-buildpack-python.git#purge
# heroku config:set HEROKU=1
#psycopg2==2.6.1
#  dj-database-url==0.3.0

## Michael Fan
pyyaml
django-cors-headers
virtualenv
